small fix entity and repository

+++ src/main/java/ru/skypro/homework/repository/CommentRepository.java
@@ -3,5 +3,7 @@ package ru.skypro.homework.repository;
 import org.springframework.data.repository.CrudRepository;
 import ru.skypro.homework.entity.CommentEntity;
 
-public interface CommentRepository extends CrudRepository<CommentEntity,Long> {
+import javax.persistence.criteria.CriteriaBuilder;
+
+public interface CommentRepository extends CrudRepository<CommentEntity, Integer> {
 }
